requirements.txt
streamlit>=1.28.0
groq>=0.4.0
tavily-python>=0.2.0
python-dotenv>=1.0.0
markdown>=3.5.0
aiohttp>=3.8.0 